refactor(rust_plan): extract warm-restore + env resolvers into submodules
`crates/soldr-cli/src/rust_plan.rs` was 1212 LOC, past the 1K LOC warn
threshold. Two self-contained sections come out into companion files:

- Warm-restore-skip sentinel logic (WarmRestoreSentinel +
  evaluate_warm_restore_skip + should_skip_warm_restore +
  write_warm_restore_sentinel + warm_restore_sentinel_path + the
  WarmRestoreSkipInputs helper) → `rust_plan_warm_restore.rs`.
- Env-var driven mode / profile / backend / tar-thread resolvers
  (rust_artifact_cache_*_from_env + parse_rust_artifact_cache_tar_threads)
  → `rust_plan_env.rs`.

```
before:  rust_plan.rs                 1212 LOC
after:   rust_plan.rs                  946 LOC   ← under warn (1000)
         rust_plan_warm_restore.rs     235 LOC   ← new, under warn
         rust_plan_env.rs               81 LOC   ← new, under warn
```

Re-exports via `pub(crate) use` so external callers keep the same path.
The private env helpers move to `pub(super)` since they're now in a
submodule but only the parent uses them.

No functional change.
